fix convertback float object failed convert to double.

+++ OgarCommon/OgarCommon.WPF.LokTarMarkup/Converters/FloatToDoubleConverter.cs
@@ -17,7 +17,7 @@ namespace OgarCommon.WPF.LokTarMarkup
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return (float)value;
+            return (float)(double)value;
         }
     }
 }
